+++ kind-integer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Version 0.1.1
 
+* Export `Abs`, `GCD`, `LCM`, `Odd`, `Even`, `toPrelude`, `fromPrelude`.
+
 * Minor cabal and documentation improvements.